added express server to add env variables
