--- index.html
@@ -1,126 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="description" content="whalebot" />
-    <meta name="author" content="whalebot inc" />
-
-    <link rel="preconnect" href="https://fonts.gstatic.com" />
-    <link href="https://fonts.googleapis.com/css2?family=Rubik:wght@400;500;700&display=swap" rel="stylesheet" />
-
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css"
-        integrity="sha512-iBBXm8fW90+nuLcSKlbmrPcLa0OT92xO1BIsZ+ywDWZCvqsWgccV3gFoRBv0z+8dLJgyAHIhR35VZc2oM/gI1w=="
-        crossorigin="anonymous" />
-
-    <!-- ScrollReveal Js -->
-    <script src="https://unpkg.com/scrollreveal"></script>
-    <script src="./js/scroll-reveal.js" defer></script>
-
-    <!-- Mixpanel -->
-    <script type="text/javascript">
-        (function (f, b) {
-            if (!b.__SV) {
-                var e, g, i, h;
-                window.mixpanel = b;
-                b._i = [];
-                b.init = function (e, f, c) {
-                    function g(a, d) {
-                        var b = d.split(".");
-                        2 == b.length && ((a = a[b[0]]), (d = b[1]));
-                        a[d] = function () {
-                            a.push([d].concat(Array.prototype.slice.call(arguments, 0)));
-                        };
-                    }
-                    var a = b;
-                    "undefined" !== typeof c ? (a = b[c] = []) : (c = "mixpanel");
-                    a.people = a.people || [];
-                    a.toString = function (a) {
-                        var d = "mixpanel";
-                        "mixpanel" !== c && (d += "." + c);
-                        a || (d += " (stub)");
-                        return d;
-                    };
-                    a.people.toString = function () {
-                        return a.toString(1) + ".people (stub)";
-                    };
-                    i = "disable time_event track track_pageview track_links track_forms track_with_groups add_group set_group remove_group register register_once alias unregister identify name_tag set_config reset opt_in_tracking opt_out_tracking has_opted_in_tracking has_opted_out_tracking clear_opt_in_out_tracking start_batch_senders people.set people.set_once people.unset people.increment people.append people.union people.track_charge people.clear_charges people.delete_user people.remove".split(" ");
-                    for (h = 0; h < i.length; h++) g(a, i[h]);
-                    var j = "set set_once union unset remove delete".split(" ");
-                    a.get_group = function () {
-                        function b(c) {
-                            d[c] = function () {
-                                call2_args = arguments;
-                                call2 = [c].concat(Array.prototype.slice.call(call2_args, 0));
-                                a.push([e, call2]);
-                            };
-                        }
-                        for (var d = {}, e = ["get_group"].concat(Array.prototype.slice.call(arguments, 0)), c = 0; c < j.length; c++) b(j[c]);
-                        return d;
-                    };
-                    b._i.push([e, f, c]);
-                };
-                b.__SV = 1.2;
-                e = f.createElement("script");
-                e.type = "text/javascript";
-                e.async = !0;
-                e.src = "undefined" !== typeof MIXPANEL_CUSTOM_LIB_URL ? MIXPANEL_CUSTOM_LIB_URL : "file:" === f.location.protocol && "//cdn.mxpnl.com/libs/mixpanel-2-latest.min.js".match(/^\/\//) ? "https://cdn.mxpnl.com/libs/mixpanel-2-latest.min.js" : "//cdn.mxpnl.com/libs/mixpanel-2-latest.min.js";
-                g = f.getElementsByTagName("script")[0];
-                g.parentNode.insertBefore(e, g);
-            }
-        })(document, window.mixpanel || []);
-    </script>
-
-    <script src='./js/script.js' defer></script>
-
-    <!-- Favicon -->
-    <link rel="shortcut icon" href="./assets/favicon.png" type="image/x-icon" />
-
-    <!-- Styles -->
-    <link rel="stylesheet" href="./css/main.css" />
-
-    <title>Whalebot</title>
-</head>
-
-<body>
-    <div class="main-container">
-        <header class="main-header">
-            <div class="header-container">
-                <nav class="main-header-navbar">
-                    <img src="./assets/fiii-removebg-preview.png" alt="CRAPPO logo" class="main-header-navbar__logo" />
-
-                    <div class="main-header-navbar__nav__item">
-                        <a href="espanol.html" class="main-header-navbar__nav__link">Espanol</a>
-                    </div>
-
-                    <div class="main-header-navbar__nav__item" id='join-telegram-btn'>
-                        <a href="https://t.me/+6ENBB7N9LZ1mMGMx" class="main-header-navbar__login__register-btn">Join telegram</a>
-                    </div>
-                </nav>
-
-                <div class="main-header-content-container">
-                    <div class="main-header-content-principal">
-                        <h1 class="main-header-content-principal__title">
-                            Profit <span>and </span> build up your Bitcoin holdings with automated onchain
-                            <span>trading on telegram</span>
-                        </h1>
-                        <p class="main-header-content-principal__description">
-                            We are currently trading in private beta and will soon be launching the pool, allowing Telegram users to join. You can join the community now on Telegram.
-                        </p>
-                        <a href="https://t.me/+6ENBB7N9LZ1mMGMx" class="main-header-content-principal__btn">Join community <i
-                                class="fas fa-angle-right"></i></a>
-                        <a href="https://x.com/whalebot0x" class="main-header-content-principal__btn">X <i
-                                class="fas fa-angle-right"></i></a>
-                    </div>
-                    <img src="./assets/header-assets/header-Illustration-modified.png" alt="" class="
-                main-header-content-principal__illustration" />
-                </div>
-            </div>
-        </header>
-        <footer>
-            <p>&copy; Whalebot 2024. All rights reserved.</p>
-        </footer>
-    </div>
-</body>
-</html>
