--- js/scroll-reveal.js
@@ -1,24 +1,0 @@
-// *******************************
-//      ScrollReveal animations
-// *******************************
-
-// -----  Main header -----
-ScrollReveal().reveal(".main-header-navbar", {
-  duration: 2000,
-  delay: 500,
-  origin: "top",
-  distance: "2rem",
-});
-ScrollReveal().reveal(".main-header-content-principal", {
-  duration: 2000,
-  delay: 500,
-  origin: "left",
-  distance: "2rem",
-});
-ScrollReveal().reveal(".main-header-content-principal__illustration", {
-  duration: 2000,
-  delay: 500,
-  origin: "right",
-  distance: "2rem",
-});
-
