+++ public/assets/features-section-assets/features-section-pattern1.png
