+++ public/index.html
@@ -1,6 +1,5 @@
 <!DOCTYPE html>
 <html lang="en">
-
 <head>
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
@@ -19,16 +18,8 @@
     <script src="https://unpkg.com/scrollreveal"></script>
     <script src="./js/scroll-reveal.js" defer></script>
 
-    <!-- Favicon -->
-    <link rel="shortcut icon" href="./assets/favicon.png" type="image/x-icon" />
-
-    <!-- Styles -->
-    <link rel="stylesheet" href="./css/main.css" />
-
     <!-- Mixpanel -->
-
-     <!-- Mixpanel -->
-     <script type="text/javascript">
+    <script type="text/javascript">
         (function (f, b) {
             if (!b.__SV) {
                 var e, g, i, h;
@@ -81,8 +72,16 @@
         })(document, window.mixpanel || []);
     </script>
 
-    <script src='/js/espanol.js'></script>
-    <title>Cryptocurrency company</title>
+    <script src='./js/script.js' defer></script>
+    <script src="/config.js"></script>
+
+    <!-- Favicon -->
+    <link rel="shortcut icon" href="./assets/favicon.png" type="image/x-icon" />
+
+    <!-- Styles -->
+    <link rel="stylesheet" href="./css/main.css" />
+
+    <title>Whalebot</title>
 </head>
 
 <body>
@@ -90,47 +89,30 @@
         <header class="main-header">
             <div class="header-container">
                 <nav class="main-header-navbar">
-                    <img src="./assets/fiii-removebg-preview.png" alt="CRAPPO logo" class="main-header-navbar__logo"  />
+                    <img src="./assets/fiii-removebg-preview.png" alt="CRAPPO logo" class="main-header-navbar__logo" />
 
-                    <!-- <ul class="main-header-navbar__nav">
-                         <li class="main-header-navbar__nav__item">
-                            <a href="#products" class="main-header-navbar__nav__link"></a>
-                        </li>
-                        <li class="main-header-navbar__nav__item">
-                            <a href="#features" class="main-header-navbar__nav__link"></a>
-                        </li>
-                        <li class="main-header-navbar__nav__item">
-                            <a href="#about" class="main-header-navbar__nav__link"></a>
-                        </li>
-                        <div class="main-header-navbar__nav__item">
-                            <a href="whalbote.finance/espanol" class="main-header-navbar__nav__link">Espanol</a>
-                        </div>
-                    </ul> -->
                     <div class="main-header-navbar__nav__item">
-                        <a href="/" class="main-header-navbar__nav__link">English</a>
+                        <a href="espanol.html" class="main-header-navbar__nav__link">Espanol</a>
                     </div>
 
-                    <div class="main-header-navbar__nav__item" id='join-telegram-btn-spanish'>
-                        <a href="https://t.me/+6ENBB7N9LZ1mMGMx" class="main-header-navbar__login__register-btn"> unirte en telegram</a>
+                    <div class="main-header-navbar__nav__item" id='join-telegram-btn'>
+                        <a href="https://t.me/+6ENBB7N9LZ1mMGMx" class="main-header-navbar__login__register-btn">Join telegram</a>
                     </div>
                 </nav>
 
                 <div class="main-header-content-container">
                     <div class="main-header-content-principal">
-                        <!-- <div class="discount-section">
-                            <span class="discount-section__discount">75% SAVE</span>
-                            <span class="discount-section__description">For the Black Friday weekend</span>
-                        </div> -->
                         <h1 class="main-header-content-principal__title">
-                            Obtén ganancias <span>y </span> aumenta tus tenencias de Bitcoin automatizando el trading onchain <span>en Telegram</span>
+                            Profit <span>and </span> build up your Bitcoin holdings with automated onchain
+                            <span>trading on telegram</span>
                         </h1>
                         <p class="main-header-content-principal__description">
-                            Actualmente estamos operando en beta privada y pronto lanzaremos el pool, permitiendo a los usuarios de Telegram unirse. Puedes unirte a la comunidad ahora en Telegram.                               
-                            <br> <br/>
-                            <a href="https://t.me/+6ENBB7N9LZ1mMGMx" class="main-header-content-principal__btn">Únete a la comunidad <i
-                                    class="fas fa-angle-right"></i></a>
-                                    <a href="https://x.com/whalebot0x" class="main-header-content-principal__btn">X <i
-                                        class="fas fa-angle-right"></i></a>
+                            We are currently trading in private beta and will soon be launching the pool, allowing Telegram users to join. You can join the community now on Telegram.
+                        </p>
+                        <a href="https://t.me/+6ENBB7N9LZ1mMGMx" class="main-header-content-principal__btn">Join community <i
+                                class="fas fa-angle-right"></i></a>
+                        <a href="https://x.com/whalebot0x" class="main-header-content-principal__btn">X <i
+                                class="fas fa-angle-right"></i></a>
                     </div>
                     <img src="./assets/header-assets/header-Illustration-modified.png" alt="" class="
                 main-header-content-principal__illustration" />
@@ -138,9 +120,8 @@
             </div>
         </header>
         <footer>
-            <p>&copy; Whalebot 2024. Todos los derechos reservados.</p>
-          </footer> 
+            <p>&copy; Whalebot 2024. All rights reserved.</p>
+        </footer>
     </div>
 </body>
-
 </html>
